chore: separate shopping cart item into its own file
Fixes https://github.com/strongloop/loopback4-example-shopping/issues/15
