+++ src/models/index.ts
@@ -4,5 +4,6 @@
 // License text available at https://opensource.org/licenses/MIT
 
 export * from './user.model';
+export * from './shopping-cart-item.model';
 export * from './shopping-cart.model';
 export * from './order.model';
